requirements.txt
crowdstrike-falconpy==1.4.1
